+++ tools/notEmpty.ts
@@ -5,6 +5,6 @@
  * @param {any} p
  * @returns {boolean}
  */
-export function notEmpty(p): boolean {
+export function notEmpty(p: any): boolean {
 	return ['', null, 'undefined', undefined].indexOf(p) < 0;
 }