+++ HISTORY.md
@@ -1,10 +1,15 @@
 ## [Unreleased]
+
+## [0.2.0] - 2019-08-05
 ### Fixed
 - cfg property of MongoApp was not correct
 ### Changed
 - Config.master renamed to app
+- `ConfigLoader` now supports loading different configs by process args
 ### Added
 - Config.log.devel
+- `Logger` class
+- `RandomTable` class
 ### Removed
 - Config.worker